requirements.txt
# FastAPI and Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM and AI
openai==1.10.0
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.16

# Vector Database
chromadb==0.4.22

# PDF Processing
pypdf==4.0.1
pdfplumber==0.10.3

# Embeddings
sentence-transformers==2.3.1

# Web Search
tavily-python==0.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Logging
loguru==0.7.2
